Configuring pure ARMA estimation
We changed our functions to allow the estimation of pure ARMA models.
We still need to treat the estimation with other algorithms, but it
seems to work quite with our comparisons with fGarch results. We may
take another look at the filter.Ama function to make sure it was
implemented correctly. The next step is to configure the other
algorithms to estimate pure ARMA and then to configure the restriction
function for stationarity.