Fit functions accepts both algorithms
1) The fit function can now accepts both ‘sqp’ and ‘nlminb’ algorithms
to perform optimisation of any model, including pure ARMA models. We
started creating the function that will calculate the stationarity
conditions of APARCH models with conditional “stable”, “get” or
“normal” distribution.

2) The next step is to test this function a little bit better using the
inequalities provided on Mittinik et al (2002) for the particular
cases. Then we will make an interface to use this function inside our
“Fit” function.

3) We also changed the Spec function to create ARMA-GARCH or
ARMA-APARCH models, instead of AR, MA or ARCH models.