t3 Distribution defined in Paolella (1997) added
We finished testing the density function. The computation of density
points can be done very fast so that MLE techniques can be done
efficiently.
We still need to include test cases for the distribution function.
In the future we plan to implement the quantile function and the random
number generator for the t3 distribution.