Included t3 distribution to the routine
The next step is to test the implementation of the rt3 function.
We also added a text file called HowToChangeThisLilbrary. This file
contains important notes and instructions for anyone that needs to
change functions or parameters inside this library, such as: adding a
new distribution function, adding a new algorithm and so on.